Allow using -d experimental in package list.
Previously, there was a "sid-list" file, as I thought we'd be also using
an "experimental-list" where to backport from. In fact now, we can use
"-d experimental" when describing a package. So this patch renames the
"sid-list" into a more generic "packages-list", and use that. The
loops around suites are also removed.

This patch also gets webkit2gtk and qpid-proton from official
jessie-backports.

Change-Id: Iba64402babf9cfcc725ee1af65a6e3ee106099bc